Add GH workflow to build and push master container to registry
Build and push container to Quay.io container registry every day. Also make it
possible to manually dispatch this action.

This container should be tested before pushing to the registry but that would
be follow up when this is working.

The container will be then used as source to run unit-tests GH action.