Fix testing target branch instead of PR branch
Hardening done in the commit baafcef39acbf0df181dd4b87176308c07a261e8 work more
than expected. Right now we are testing target branch instead of the PR ref
which means that we are not testing commits on the PR at all.
Ups...

We have to change what we are checking out to fix this.