workflows: Add manual triggering of tests
Run unit tests in a PR triggered by a "/test [...]" comment from an
organization member.  This avoids running untrusted and unreviewed code
on self-hosted runners.

For now this only applies for PRs against the rhel-8 branch, but an
`issue_comment` trigger needs to be on master. (In the future we *could*
support running rhel/ELN tests on master as well, though.)