Fix localization issues in GUI & TUI
Fix localization context issues in GUI by direct string translations
instead of just marking strings as translatable in spokes and hubs
and translating them before use. This is fragile as strings might
be marked in Initial Setup or Anaconda Addons yet translated and
displayed in Anaconda, with context that does not match.

We can't do this directly in class attributes due to these being
evaluated much to early. Instead, put strings into properties and
static methods. Also make sure these are implemented by marking the
properties and static methods as abstract methods.

When we are at it convert some non-localization class attributes
to properties and static methods for better consistency.

While doing all this, there is one more thing we need to handle,
spoke ordering. So far spokes have been ordered in their respective
categories by alphabetical sorting based on their untranslated English
name. This was simple and reasonably stable (spokes would not change
position unless they were renamed). With the change to direct runtime
translation, this can no longer be done as only the translated spoke
name is available.

Rather than adding yet another property or static method, we decided
to introduce the get_sort_order() static method, that unambiguously
defines the order of spokes in a category. We already have a similar
mechanism for categories. This way we make the spoke ordering
independent of spoke names, resolving the potential issues with
language specific ordering. And while we attempt to keep the existing
spoke order by setting the sort order values accordingly, we now have
the option to change the ordering as needed in the future.