Drop anaconda-ci container entrypoint
It's rather pointless to copy entrypoint.sh into the container: Without
the bind mount, running the container will just fail (as there is
nothing to build), and with the bind mount the entrypoint can and should
be taken from the bind-mounted source instead of the static copy in the
container, so that it's much easier to test changes to the script.

 - Drop ENTRYPOINT and CMD from the container, so that by default you
   just get a shell.

 - Adjust the `make container-ci` rule to dynamically supply the
   entrypoint and test command.

 - As the above breaks the documented "run with interactive shell"
   command, provide a `make container-shell` for symmetry and
   convenience. This then covers both styles of working with the
   container (throw-away ci run or interactive debugging).

Cherry-picked from master commit 8d5d62d37c4.

Related: rhbz#1885635