Don't clobber host checkout on `make container-ci`
Copy the work tree into the container's file system, so that autotools,
built/pyc files etc. don't clobber the host checkout with files (which
are not even owned by the user, so are a bit difficult to delete).

Just copy out the logs to the host's ./test-logs/*.log directory. Adjust
the validate workflow accordingly.