Change handling of UID and GID values
Add separate field to UID and GID for (un)set state: It was not entirely safe,
correct and future-proof to mix the numerical value and signalling state/mode
in one numerical variable. Thus adding a separate status field to communicate
whether the value is set or not.

Further, add methods to simplify handling UIDs and GIDs. These methods are
added to the D-Bus structures for users and groups to convert the
number + mode pair to a value that can be an int or None, and back. This
simplifies again most handling of these values after the number and mode
split.

Related: rhbz#1786680