Select disks for implicit partitions (#1642391)
Don't use disks that will be used for allocation of requested partitions
(for example, the /boot partition). Calculate the number of slots that
will be used by requested partitions and skip disks that will not have
free slots for implicit partitions.

Otherwise, we might schedule implicit partitions, that wouldn't be possible
to allocate on the specified disk, and the partitioning might fail.

Resolves: rhbz#1642391