Fix user permission checking for rhel-8 validate task
It does not work correctly for users with private visibility on the
organization. However, this is problem because the private visibility is default
on GitHub and it can't be changed globally.

Use GH API to check user permissions on the repository. This does not require
special privileges on token. Allow only admin or write access to start the tests.