Add rhel unit-tests workflow for pull_request_trigger back
This is simplification for users with write access to have tests started even
without the comment so it will work as expected.

These tests will be no-op for external contributors. For them only first job is
executed and this job only getting information but not testing anything.

This workflow file is taken from master branch where it runs on comment. We did
a few modifications to it:
- use pull_request_target trigger so it will be automatically called when code
  is updated
- rename workflow to avoid collisions and make it crystal clear that this will
  run only for owners
- cleanup unnecessary code; we don't need any API calls except checking for user
  privileges everything else could be read from the trigger event data directly

Related: rhbz#1885635