Return the exit status of a Python module that failed to start
If the Python module started by the start-module script fails before the
timeout of sixty seconds, return its exit status. Otherwise, return zero.

This will speed up the propagation of the error to the boss. Otherwise,
we wait for the timeout of service_start_timeout defined by the Anaconda
message bus.

The start-module script runs in a separate process and doesn't block the
message bus, the boss or the installer.