Run master tests on Fedora ELN
Let's make sure that Anaconda keeps building/working on ELN, to avoid
breakage in RHEL 9.

Add an `eln` argument to the anaconda-ci Dockerfile, and conditionally
build ELN instead of Rawhide if set. There is no official Fedora ELN
container image yet, so install the ELN repo file manually and
cross-grade from Rawhide.

For PRs against master, build an ELN container in the GitHub workflow
and run tests against it in an additional job. Once we get the ELN
container into daily builds and published on quay.io, the build should
be conditionalized like for the regular anaconda-ci container.

Add a `set -e` to the Dockerfile's main `RUN` step, so that `if`
conditions and multiple lines become a bit less unwieldy (no need to
write `&&` everywhere).