Skip add-on modules that failed to start
If the add-on fails to start, just remove it from the list of available modules
and continue. Don't abort the whole installation because of that.

This fix will allow the kdump add-on to refuse to run its DBus module if it is
not enabled via the inst.kdump_addon boot option.