Disable seccomp profile when running containers on Ubuntu host
Latest Fedora Rawhide has a toolchain change which renders bash's
builtin `test -r` command incompatible with Ubuntu's container seccomp
profile (which applies to both docker and podman). This leads to
libtool, autoconf, and other programs to fall over trying to find `m4`
and other programs.

Add the option in our Makefile instead of the github workflow for the
anaconda-ci container, so that it also works when running interactively
on a Ubuntu host (i.e. pretty much every cloud instance, like Travis or
Semaphore). This also avoids repetition between the validate and
container-autoupdate workflows.