Set shorter workflow timeouts
The default timeout is 6 hours, which just causes unnecessary hangs,
delays, and blocked self-hosted runners if anything goes wrong (such as
https://github.com/PyCQA/pylint/issues/3899).

A normal run takes about 10 minutes, so set a timeout of 30 minutes.
Use one hour for the full container rebuild+validation+upload to be on
the safe side (it usually takes between 15 and 18 minutes).