Fixup category sorting
Turns out the category sorting code is wrong
(as categories are classes, not class instances :P)
yet it still manages to provide the expected output,
so this incorrect behavior is not readily apparent.

Thanks vponcova for spotting this! :)

Also when we are at it drop apparently unnecessary sorting
of categories in the collect function and fixup the unit test.
Make sure the input is not sorted in the expected order and
move the test to the appropriate sub folder.