Lower geolocation connection timeout to 5 seconds (#1774873)
Lower the connection timeout to 5 seconds as the GUI will wait
only up to 3 seconds for the geolocation thread to finish anyway.

Up till now we used the Anaconda default timeout, which is 45 seconds.
Apparently the actual waiting time could be up to 2x the timeout value,
so 45 seconds was far too long.

Resolves: rhbz#1774873