Drop old WiFi & geocoding based geolocation support
This code was developed in the initial stages of geolocation
support in Anaconda with the aim of providing an alternative
to purely IP based territory & timezone determination.

In was considered a prototype back then and was never documented as
something you can enable via the inst.geoloc boot option.

In the end we never did get to making the code production ready
while the IP based geolocation worked fine. Also all the WiFi
location and geocoding API URLs no longer work anyway.

So lets drop that old code & cleanup the docstrings a bit when we are at it.