Stop trying to set the default hostname
If the user configured a hostname, then great, let's use it. If not,
let's just do nothing. This is totally appropriate, we know that
systemd sets some hostname when the machine is booted and we can just
keep using it.

In https://github.com/rhinstaller/anaconda/pull/643:
> To prevent fallouts from changed behaviour, we keep writing out
> /etc/hostname even when static hostname is not configured - using
> value "localhost.localdomain". In this case target system still
> applies transient hostname (same as if we didn't write
> /etc/hostname). I bet there is something relying on /etc/hostname
> existing after installation.

I want to drop this behaviour from systemd since it's just pointless
complicated. All the layers of guesswork mean that the whole system is
frozen in place. Instead of trying to second-guess the user and figure
out if the hostname was configured by the user or not based on the
specific string used, let's just write the config file only when the
user specifies a hostname.