Build real boot.iso for /kickstart-tests runs
Using an updates.img is really fast, but does not catch changes to
dracut or file removals. So move to building a boot.iso from Rawhide and
locally built anaconda RPMs, and using that.

Do the lorax build *after* checking out kickstart-tests and starting the
squid container, so that lorax already benefits from the HTTP caching
and seeds the cache for the subsequent kickstart tests.

Keep the old updates.img build in a comment, as that may come in handy
in the future -- we may want to control that via the magic comment to
select between fast and thorough?