Add support to run RPM tests in a container
This new make command will do this:

- build RPM files in a container
- run rpm-tests
- test installation of these RPM files

This commit is extending run-build-and-arg script because it's not reasonably
possible to put multiple commands from Makefile in the podman. To avoid this
issue it's adding --build-rpms as first argument which when detected will build
RPM and test installation before the command execution.

Also fix issues that nosetests-3 executable is not available in the container.

Related: rhbz#1885635