Define container for running rpm-tests
Define and document the  "anaconda-rpm" container similar to the
"anaconda-ci" one for running the RPM tests. We don't want to use the
anaconda-ci one as that has too many packages pre-installed, and thus we
would not detect missing dependencies.

Fix the workflow to rebase the branch to current master just as the
unit tests do.