Remove the error handler for the missing image
We should fail to set up the source and report that to the user instead.

The mountImage function is called only from the _find_and_mount_iso method with
a normalized path of an ISO file. Therefore, we can remove the mountImage
function and use the mount function from the payload utils instead.