Deprecate the interactive-defaults.ks file
This file is deprecated and will be removed in in the future. Use the Anaconda
configuration files instead.

Anaconda no longer uses this file for the configuration of the interactive
defaults, but some of the products do. Make sure that the products use the
configuration files before the removal of this support.