GH workflow: do not fail all matrix jobs on failure of one
The fail-fast configuration value is used to kill all jobs in a matrix when
first fails. However, we don't want to do that. These jobs are not dependent and
we want to know if one arch is able to run when other is not.

Change this for validate and container-refresh workflows.