Handle correctly too slow X11 start
Handling of the SIGUSR1 signal after X11 start timed out was wrong. Normally,
this signal is used to test the exception handler. If X11 timed out but
eventually started and sent SIGUSR1 to Anaconda, the signal was interpreted as
testing the exception handler and Anaconda crashed.

With this change, the X server is terminated to ensure it won't keep starting
and send the signal. The SIGUSR1 handler is restored only after the
termination is finished, so there is little chance of misnterpreting the
signal.

This also changes the failure mode. Now, it is the X11 termination that can
take long or forever.

Resolves: rhbz#1918702

Co-authored-by: bitcoffee <854182924@qq.com>
Co-authored-by: Vendula Poncova <vponcova@redhat.com>