anaconda should add only s390utils-core
The s390utils-core is the sufficient part of s390utils that's required for a s390x
system to be functional. Thus add only s390utils-core to the package set instead of
the complete s390utils. Keep the requirement for the bootloader, but omit it for
the platform, where requirements are set via comps.

Resolves: rhbz#1918642