Fix root password and LUKS passphrase visibility toggle (#1911360)
Anaconda supports showing a little "eye" icon in the password/passphrase
entry field that can be used to make the entered text visible in
plaintext during entry.

For this to work correctly entry text visibility state needs to be
set correctly. This has been broken by some of the password/passphrase
entry related refactorings in the recent past, resulting in the "eye"
icon not being visible.

So put the initialization back for the root password and LUKS passphrase
entries and also extend the docstring in the user configuration spoke
to note how password visibility influences the toggle icon.

Resolves: rhbz#1911360