Modify spec file to add patches required on rhel-9 rebuild (#1907566)
We made the strong requirement for the 'inst.' prefix for Anaconda kernel boot
arguments. However, unfortunately beaker needs more time to adapt and we are
forced to merge the changes because of Fedora System Wide Change Completion
Deadline. So these patches will enable beaker to work for RHEL-9 for the
required time. See the attached bug for more info.

Related: rhbz#1907566