Fix base container specification in refresh workflow
We should use BASE_CONTAINER instead of CONTAINER_BUILD_ARGS because now the
build args are passed to the build in the Makefile directly.

This worked even before because we basically re-defined value from before but it
wasn't clean. Now we have clean solution.

However, we can't use environment variable otherwise it will be empty instead of
having default value. So use that only if it's defined.