Use older Ubuntu for container auto-update workflow
We need older Ubuntu to avoid a bug which will make podman run on ELN image
fail. Thanks to that we can avoid using Docker which was not that easy to be
used.

Older version seems to have older crun which doesn't have this issue.

See: https://github.com/containers/podman/issues/9442

Suggested-by: Vladimir Slavik <vslavik@redhat.com>