Determine GRUB directory relative path to use in config file
Currently Anaconda is only checking whether /boot/ is a mount point or not
to use the proper relative path to the GRUB directory. But there are more
cases than this, for example /boot/grub2 could be a symlink or a subvolume
in a btrfs filesystem.

Use the grub2-mkrelpath tool to figure out the actual relative path that
has to be used in the GRUB config file.

Resolves: rhbz#1928588

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>