Add ParseOrganizationData runtime DBus task
Add a new runtime DBus task that can query information about the
organizations of a Red Hat account.

This information is not just important to find out if the user is a
member of multiple organizations, but also is the organization the user
is a member of has Simple Content Access enabled or not.

The date is obtained by calling the GetOrgs() method of the RHSM Dbus API
in JSON form and is parsed into a list of OrganizationData DBus structs,
one per organization.

At the moment we only parse and use organization name, id and
entitlement status, but this can be easily extended in the future if we
need to use more of the available data for some purpose.

Related: rhbz#1921131