ostree: ignore exit code 65 for systemd-tmpfiles
When pre-filling /var after deploying the ostree commit, do not
interpret return code 65 by systemd-tmpfiles as an error. This
return code is used to indicate that some of the lines in the
configuration were ignore, which could be due to a syntax error
but also something as harmless as duplicated entires. Either way
it the contents of `/lib/tmpfiles.d` is provided the content and
thus out of the hand of the installer itself. Since there was no
error executing the not-ignored lines we should not abort the
installation.

Resolves: rhbz#1935648