Remove crun version abort for container refresh workflow
This abort logic works great, however, GitHub runners are not updated
immediately. That means that this check will abort the workflow for some runners
but and the others will work.

Lets remove this and just remove the workaround as whole when we see that the
old runners are not available anymore.

Also switch to ubuntu explicit version because latest tag depends on runner you
will get. It could be 18.04 or 20.04 based on documentation.