Replace use of imp with importlib in collect()
The imp module has been deprecated with Python 3.4.

The new solution does not need the finally clause because importlib does not
open an user-facing file descriptor with the search.

File extensions are kept with the leading dot.

At the moment of writing this, the manual loading via importlib is needed for
Initial Setup EULA spoke, and probably for the addons.