Check source type directly, not via DBus
Not all payloads are available via DBus in RHEL 8 Anaconda,
so use the payload instance directly when checking payload type
in the Connect to Red Hat spokes. Also in some cases the payload
type could be None instead of a string, so handle that as well.

This should fix the issue with RHV, which uses the liveimg payload,
which is not available via DBus on RHEL 8, leading to a crash.

Resolves: rhbz#1941578