Check source type directly, not via DBus
Not all payloads are available via DBus in RHEL 9 Anaconda,
so use the payload instance directly when checking payload type
in the Connect to Red Hat spokes. Also in some cases the payload
type could be None instead of a string, so handle that as well.

This should fix the issue with RHV, which uses the liveimg payload,
which is not available via DBus on RHEL 9, leading to a crash.

(cherry picked from commit d6ea3b39b839e4822ea7d1e9afea5fb81352b732)

Related: rhbz#1941578
Resolves: rhbz#1955311