Replace our Fedora owners check by GH environments for our workflows
This will enable us to simplify our current solution a lot. We can merge
contributors and owners workflow and avoid using REST API to set statuses etc...
Finally we don't have to change everything on two places!

This has one drawback. By doing this we will gate everything, including our PRs.
Unfortunately, I wasn't able to find a way how to enable this just for
contributors but I think it's worth the effort because we can approve the change
ourselves with just 3 clicks.