Use GitHub environment gating for kstests workflow
Remove our custom solution to verify user but instead leave this to environment.
That means that we don't have to care and mainly we don't have to test if we
our changes are not breaking this.

The problem is, that we always needs to authorize the run, even for us. It will
work the same way for contributors as for us. That is a bit unfortunate but I
think it's a worth the gain.

However, thanks to the solution that kickstart tests are started on comment so
the workflow run is not attached to the PR we have to tell user where they can
approve the run. Let's do that by commenting on the PR with link and request for
approval.