[payload]Adds priority and modules_hotfixes params
This commit follows the PR pykickstart/pykickstart#378

As cost is already part of the supported parameter, this commit just adds the support for priority which could be useful when using repositories that could provide the same packages.

And module_hotfixes may be not as useful but it allows to supports hotfixes repositories.
Thoses repositories could be used in cases of creating images regurlarly in a CI system.

Change-Id: If9a3992fa4e0db794485bd689857c8574b5e180f
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>