subscription: allow dates in ISO 8601 format
Right now, the date returned by D-Bus method calls of
subscription-manager have a locale-specific format, which can be hard to
parse. A planned change is to return all the dates in ISO 8601 format:
https://github.com/candlepin/subscription-manager/pull/2551

Hence, make the subscription module able to cope with the future value:
first try to parse the date as ISO 8601, and if it does not work try to
parse as it is done today. Add small unit testing data to check that
both ISO 8601 and the locale-specific formats work.