Add a shim script to translate signals from Xorg
Handling of the SIGUSR1 signal after X11 start timed out was wrong. Normally,
this signal is used to test the exception handler. If X11 timed out but
eventually started and sent SIGUSR1 to Anaconda, the signal was interpreted as
testing the exception handler and Anaconda crashed.

With this change, a shim script translates the signals to make sure they are
no longer the same: SIGUSR1 - > SIGRTMIN+8 (most probably 42).

A non-responsive Xorg is left to be.

Resolves: rhbz#1918702

Co-authored-by: bitcoffee <854182924@qq.com>
Co-authored-by: Vendula Poncova <vponcova@redhat.com>