Remove parse-kickstart tests which don't test anything useful
The tests became senseless after generating of ifcfg files and its
testing was removed from initramfs.

I will add the tests when/if support for activation in initramfs from
kickstart configuration for these cases (team, vlan) is added or at
least parse-kickstart is fixed to not generate wrong dracut arguments in
the cases.