Handle SIGUSR1 correctly after Xorg timeout
When Anaconda receives the SIGUSR1 signal, it executes its crash handler. When
Anaconda launches Xorg and it successfully starts, it sends SIGUSR1 to
Anaconda. Thus, we use a special handler for this signal while waiting for
Xorg to start, and a "normal" one otherwise.

Previously, if a Xorg start timed out, the temporary Xorg start handler was
always replaced with the handler that started exception handler test. However,
if Xorg kept starting in the background, it would eventually send that signal,
which would be interpreted as a test of the exception handler. Then, Anaconda
would stop with the error report message.

This changes the handling of SIGUSR1 after Xorg start times out, such that it
is no longer wrongly interpreted as a QE test, but correctly as a late Xorg
start. To prevent the newly started Xorg from stealing the screen, receiving
this signal switches back to tty1 with tmux and TUI.

Resolves: rhbz#1918702
Resolves: rhbz#1955306

(cherry picked from commit 4434c32bb86627b96198f2e20d41e643e062f776)