Remove comments about ConditionArchitecture
We do not want to use it, the current solution does exactly what we want.

Details from
https://github.com/rhinstaller/anaconda/pull/3327#issuecomment-831310585
and below, as posted by jstodola from QE/RTT:

The problem is that /sys/hypervisor/s390 exists when the system is running as
a z/VM guest, but it doesn't exist when running as a KVM s390x guest.
ConditionArchitecture=|s390x will be true for both z/VM and KVM, resulting in
sshd/anaconda-direct being executed by default on s390x KVM guests.

The reason for running sshd by default on z/VM guests or in LPAR is that the
console is not very user friendly and for example scrolling in the console
can be done by whole pages, not by lines (I'm talking about x3270, as one of
the recommended program for connecting to z/VM guests). Also the default
resolution 80x43 characters is not too much for displaying anaconda in text
mode. KVM s390x guests on the other hand behave mostly like x86_64 KVM guests,
with usable console and even a graphical adapter. So, yes, it's expected that
both virtualization methods behave differently.