Drop the "Allow SSH root login with password" checkbox (#1855736)
The "Allow SSH root login with password" checkbox in the root password
spoke was added as a temporary migration helper for user when Fedora
disable password based SSH login for the root account back in 2019.
This in itself was actually just restoring OpenSSH upstream default
behavior, as the OpenSSH upstream disabled password based root login
already back in 2015.

As enough time has been given for migration to more secure remote access methods
that are less prone to remote brute force password guessing (such as key based
authentication or user account with unique username) it is now time to
drop the temporary checkbox from the Anaconda GUI.

Resolves: rhbz#1855736