Do not use cache when building containers in GitHub workflow (#infra)
We are setting CONTAINER_BUILD_ARGS env variable in Fedora which was also
brought here. The problem is that CONTAINER_BUILD_ARGS are also Makefile
variable so instead of having the default value which is "--no-cache" we will
set an empty string which will enable the cache back.

We don't use this env variable on RHEL so it's safe for removal.