Do not use cache when building containers in tests (#infra)
The cache during the containers build is not reliable because it cannot detect
changes in most use-cases. Let's avoid using it to be sure.

The `--no-cache` is default when executing Makefile build commands, however,
thanks to the environment variable, we are creating here, we erase the original
Makefile value. This was introduced to change build arguments for ELN builds
but when no arguments are used it will set "" and just erase the --no-cache
default argument. Let's add it back workflow all the time.