Remove base container from kickstart-test workflow (#infra)
We don't need to specify base container anymore. When we use container instead
of doing the build manually it's resolved based on the anaconda branch-config.mk
settings when Makefile is invoked. The only place in the future were we need
this could be ELN but we don't support that now.