Add check for infrastructure PRs (#infra)
This check will verify that PR is changing only files for the infrastructure and
won't touch files important by the project.

This check will be started for all PRs marked by the 'infrastructure' label.

Add configuration file to the project which could be used for future project
structure. Could be used for new updates_image generation.