Apply the bootloader options before the installation
Some of the bootloader options are not needed until the installation.
If we apply them during the scheduling of the partitions, it won't be
possible to change them after that, for example, from an add-on. Let's
apply these options later, right before the bootloader installation.

(cherry-picked from a commit dbac59a)

Related: RHEL-17205