Use secrets from the environment not organization (#infra)
We are possible to remove gating on Rawhide by using GH permissions. That will
allow us to reduce what user can do with the repository. However, we can't solve
access to secrets this way. To avoid users reading our secrets we will move the
secrets from organization level to environment which will be used only on place
where it should be used.