Add Satellite support to asynchronous registration handler
Add Satellite support to the asynchronous registration/unregistration
that is run for both interactive and kickstart installations.

This includes moving parts of the register-and-subscribe functionality
to the RegisterAndSubscribe task that also includes the Satellite bits
needed at registration time.

Ideally we would like to move *all* the logic to a scheduling task,
but this is only realistically possible once payload has been modularized
and can be fully controlled via DBus.

Resolves: rhbz#1951709