Add new configuration options for the DBus module activation
Use the activatable_modules option to specify Anaconda DBus modules that can be
activated. Use the forbidden_modules option to specify modules that are not
allowed to run. Use the optional_modules to specify modules that can fail to
run without aborting the installation.

The addons_enabled and kickstart_modules options are deprecated and will be
removed in the future.

Related: rhbz#1957063