Specify COPR chroot to use in our Dockerfiles (#infra)
The Fedora COPR has some issue to detect version of the OS in the Fedora
Rawhide container to choose which chroot to use. That resulted in a failed build
of the container. Let's specify this in the Dockerfiles.

This solution is designed to work also on forked Anaconda branches for to be
released Fedoras. In that case the branch will be fXX-(devel|release) so just
remove everything from '-' symbol.