Add new error reporting hook when Dracut timeout (#1983098)
This will add a new message during the Dracut timeout cycle and the cycle will
end with the message on screen waiting user to decide if they want to access
terminal.

The message is focusing on missing inst.stage2/inst.repo parameter. This is just
a hint because there are more reasons why the boot could not be successful.
Still this is the probable reason. Especially after we require inst. prefix for
our kernel boot options.

Resolves: rhbz#1983098