Add root_password_must_be_set config option
Add the root_password_must_be_set Anaconda config file option that
requires root password to be set during the given installation run.

This option works for both TUI and GUI and overrides the normal behavior
where setting an admin user is sufficient and where root password
configuration is not enforced for kickstart installation at all.

As this feature was requested by RHV, also enable it in the RHV and
oVirt configuration files when we are at it.

Resolves: rhbz#1981807