Fix exception message tests which were used incorrectly (#infra)
On a few places in the code we used:

self.assertRaises(Exception, msg="message of exception"):

However, this is not a valid usage. The `msg` argument here is replacement of
the error message raised from the assert, not a message of the exception to
compare.

The conversion script to pytest dropped these `msg` arguments (probably because
pytest doesn't support them). So I'm fixing these tests now.